Informs retroarch that PNG extension is also supported (#2832)
* Informs retroarch that PNG extension is also supported

* Loading PNG cart from cart.c (fix PNG load in libretro)

Moved loadPngCart from studio.c to cart.c
The standalone binary loads carts from this stack:
studio.c: studio_create()
console.c: initConsole()
console.c: cmdLoadCart()

cmdLoadCart() had a working PNG cart load,
but libretro uses `tic.c: tic80_load()` directlty, but that had a
incompete PNG load.

So I tried to combine and bring the aproach from cmdLoadCart() to
tic80_load() by letting them both use the same PNG loading function.

* libpng CMakelist - order may matter

* target_link_libraries(tic80core PRIVATE png)
