+++ cmake/core.cmake
@@ -29,6 +29,7 @@ set(TIC80CORE_SRC
     ${TIC80CORE_DIR}/ext/fft.c
     ${TIC80CORE_DIR}/ext/kiss_fft.c
     ${TIC80CORE_DIR}/ext/kiss_fftr.c
+    ${TIC80CORE_DIR}/ext/png.c
 )
 
 if(BUILD_DEPRECATED)
@@ -55,6 +56,7 @@ target_include_directories(tic80core
         ${CMAKE_SOURCE_DIR}/include
         ${CMAKE_SOURCE_DIR}/src)
 
+target_link_libraries(tic80core PRIVATE png)
 target_link_libraries(tic80core PRIVATE blipbuf)
 
 if(BUILD_WITH_ZLIB)
